Corrected errors related to BackupManagerGit class
