+++ src/main/java/org/jabref/logic/util/io/BackupFileUtil.java
@@ -9,7 +9,7 @@ import java.time.format.DateTimeFormatter;
 import java.util.HexFormat;
 import java.util.Optional;
 
-import org.jabref.gui.autosaveandbackup.BackupManager;
+import org.jabref.gui.autosaveandbackup.BackupManagerGit;
 import org.jabref.logic.util.BackupFileType;
 
 import org.slf4j.Logger;
@@ -34,7 +34,7 @@ public class BackupFileUtil {
      *     In case that fails, the return path of the .bak file is set to be next to the .bib file
      * </p>
      * <p>
-     *     Note that this backup is different from the <code>.sav</code> file generated by {@link BackupManager}
+     *     Note that this backup is different from the <code>.sav</code> file generated by {@link BackupManagerGit}
      *     (and configured in the preferences as "make backups")
      * </p>
      */
